+++ src/main/java/com/liquibaseproject/exception/ApplicationError.java
@@ -1,0 +1,11 @@
+package com.liquibaseproject.exception;
+
+import lombok.Builder;
+
+@Builder
+public class ApplicationError {
+
+    public Integer code;
+    public String type;
+    public String message;
+}