+++ src/main/java/com/liquibaseproject/exception/EmptyInputException.java
@@ -1,7 +1,0 @@
-package com.liquibaseproject.exception;
-
-public class EmptyInputException extends RuntimeException {
-    public EmptyInputException(String message) {
-        super(message);
-    }
-}