+++ src/main/java/com/liquibaseproject/service/OrdersService.java
@@ -1,7 +1,8 @@
 package com.liquibaseproject.service;
 
 import com.liquibaseproject.entity.Orders;
-import com.liquibaseproject.exception.ResultsNotFoundException;
+import com.liquibaseproject.exception.ExceptionUtil;
+import com.liquibaseproject.exception.LiquibaseProjectException;
 import com.liquibaseproject.repository.OrdersRepository;
 import io.micrometer.common.util.StringUtils;
 import org.springframework.stereotype.Service;
@@ -17,8 +18,6 @@ public class OrdersService {
 
     private final OrdersRepository ordersRepository;
 
-    private static final String ORDER_NOT_FOUND_MESSAGE = "Order with ID %s does not exist.";
-
     public OrdersService(OrdersRepository ordersRepository) {
         this.ordersRepository = ordersRepository;
     }
@@ -27,7 +26,7 @@ public class OrdersService {
         if (ordersRepository.existsById(id)) {
             return ordersRepository.findById(id);
         } else {
-            throw new ResultsNotFoundException(String.format(ORDER_NOT_FOUND_MESSAGE, id));
+            throw ExceptionUtil.logAndBuildException(LiquibaseProjectException.ORDERS_NOT_FOUND_EXCEPTION);
         }
     }
 
@@ -48,7 +47,7 @@ public class OrdersService {
 
     public void updateOrder(UUID id, Orders ordersEntity) {
         Orders order = ordersRepository.findById(id)
-                .orElseThrow(() -> new ResultsNotFoundException(String.format(ORDER_NOT_FOUND_MESSAGE, id)));
+                .orElseThrow(() -> ExceptionUtil.logAndBuildException(LiquibaseProjectException.ORDERS_NOT_FOUND_EXCEPTION));
 
         if (Objects.nonNull(ordersEntity.getQuantity())) {
             order.setQuantity(ordersEntity.getQuantity());
@@ -65,7 +64,7 @@ public class OrdersService {
         if (ordersRepository.existsById(id)) {
             ordersRepository.deleteById(id);
         } else {
-            throw new ResultsNotFoundException(String.format(ORDER_NOT_FOUND_MESSAGE, id));
+            throw ExceptionUtil.logAndBuildException(LiquibaseProjectException.ORDERS_NOT_FOUND_EXCEPTION);
         }
     }
 }