+++ src/main/java/com/liquibaseproject/service/ProductsService.java
@@ -1,7 +1,8 @@
 package com.liquibaseproject.service;
 
 import com.liquibaseproject.entity.Products;
-import com.liquibaseproject.exception.ResultsNotFoundException;
+import com.liquibaseproject.exception.ExceptionUtil;
+import com.liquibaseproject.exception.LiquibaseProjectException;
 import com.liquibaseproject.repository.ProductsRepository;
 import io.micrometer.common.util.StringUtils;
 import org.springframework.stereotype.Service;
@@ -16,8 +17,6 @@ public class ProductsService {
 
     private final ProductsRepository productsRepository;
 
-    private static final String PRODUCT_NOT_FOUND_MESSAGE = "Product with ID %s does not exist.";
-
     public ProductsService(ProductsRepository productsRepository) {
         this.productsRepository = productsRepository;
     }
@@ -26,7 +25,7 @@ public class ProductsService {
         if (productsRepository.existsById(id)) {
             return productsRepository.findById(id);
         } else {
-            throw new ResultsNotFoundException(String.format(PRODUCT_NOT_FOUND_MESSAGE, id));
+            throw ExceptionUtil.logAndBuildException(LiquibaseProjectException.PRODUCTS_NOT_FOUND_EXCEPTION);
         }
     }
 
@@ -44,7 +43,7 @@ public class ProductsService {
 
     public void updateProduct(UUID id, Products productsEntity) {
         Products product = productsRepository.findById(id)
-                .orElseThrow(() -> new ResultsNotFoundException(String.format(PRODUCT_NOT_FOUND_MESSAGE, id)));
+                .orElseThrow(() -> ExceptionUtil.logAndBuildException(LiquibaseProjectException.PRODUCTS_NOT_FOUND_EXCEPTION));
 
         if (StringUtils.isNotEmpty(productsEntity.getName())) {
             product.setName(productsEntity.getName());
@@ -69,7 +68,7 @@ public class ProductsService {
         if (productsRepository.existsById(id)) {
             productsRepository.deleteById(id);
         } else {
-            throw new ResultsNotFoundException(String.format(PRODUCT_NOT_FOUND_MESSAGE, id));
+            throw ExceptionUtil.logAndBuildException(LiquibaseProjectException.PRODUCTS_NOT_FOUND_EXCEPTION);
         }
     }
 }